+++ src/main/java/dev/jaxydog/astral/content/item/Customized.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: AGPL-3.0-or-later
  *
- * Copyright © 2024 Jaxydog
+ * Copyright © 2023–2024 Jaxydog
  *
  * This file is part of Astral.
  *
@@ -14,21 +14,53 @@
 
 package dev.jaxydog.astral.content.item;
 
-import dev.jaxydog.astral.utility.NbtUtil;
 import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NbtCompound;
 
-/** Represents an item that uses custom model data to render */
+import java.util.Optional;
+
+/**
+ * Provides common functionality for items that are expected to contain the {@value #CUSTOM_MODEL_DATA_KEY} NBT key.
+ *
+ * @author Jaxydog
+ */
 public interface Customized {
 
-    /** The NBT key that corresponds to the built-in custom model data renderer */
+    /**
+     * The NBT key corresponding to an item's custom model identifier.
+     * <p>
+     * This is directly taken from
+     * {@link net.minecraft.client.item.ModelPredicateProviderRegistry#CUSTOM_MODEL_DATA_KEY}, where it is defined as a
+     * private static constant.
+     */
+    @SuppressWarnings("JavadocReference")
     String CUSTOM_MODEL_DATA_KEY = "CustomModelData";
 
-    /** Returns the stack's custom model identifier */
-    default int getCustomModelData(ItemStack stack) {
-        return NbtUtil.getInt(stack, CUSTOM_MODEL_DATA_KEY);
+    /**
+     * Returns the given stack's assigned custom model data.
+     * <p>
+     * If the key is present but is not assigned to a valid number, {@code Optional.of(0)} is returned.
+     *
+     * @param stack The item stack.
+     *
+     * @return The custom model data.
+     */
+    default Optional<Integer> getCustomModelData(ItemStack stack) {
+        final NbtCompound compound = stack.getNbt();
+
+        if (compound == null || !compound.contains(CUSTOM_MODEL_DATA_KEY)) {
+            return Optional.empty();
+        } else {
+            return Optional.of(compound.getInt(CUSTOM_MODEL_DATA_KEY));
+        }
     }
 
-    /** Returns the stack's custom model identifier */
+    /**
+     * Sets the given stack's custom model data.
+     *
+     * @param stack The item stack.
+     * @param data The custom model data.
+     */
     default void setCustomModelData(ItemStack stack, int data) {
         stack.getOrCreateNbt().putInt(CUSTOM_MODEL_DATA_KEY, data);
     }